% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_se.R
\name{bootstrap_se}
\alias{bootstrap_se}
\title{bootstrap_se}
\usage{
bootstrap_se(...)
}
\arguments{
\item{...}{args to pass through}
}
\value{
character html fragment
}
\description{
function to use to make a custom rendering of Rmd files
}

