% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bs_modal}
\alias{bs_modal}
\title{bs_modal}
\usage{
bs_modal(id, title, trigger, ..., href)
}
\arguments{
\item{id}{}

\item{title}{}

\item{trigger}{}

\item{...}{}

\item{href}{}
}
\value{
character, html fragment
}
\description{
create html for bootstrap modal - based on ebailey78/shinyBS bsModal
}

