% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/button.R
\name{button_external}
\alias{button_external}
\alias{button_external_app}
\alias{button_external_page}
\title{button_external}
\usage{
button_external(url, label = "", class = "btn btn-primary")

button_external_page(url, label = "")

button_external_app(url, label = "")
}
\arguments{
\item{url}{character, external link e.g. "https://ijlyttle.shinyapps.io/introduction"}

\item{label}{character, html for the button}

\item{class}{character, describes the class of the button ("btn btn-primary")}
}
\value{
html fragment describing a div with a centered button
}
\description{
generate html for a button to launch
an external page in a new browser-tab
}

