% Generated by roxygen2 (4.0.2): do not edit by hand
\name{button_modal}
\alias{button_modal}
\title{button_modal}
\usage{
button_modal(id, idModal, class = "btn", icon = NULL, label = "")
}
\arguments{
\item{id}{}

\item{idModal}{}

\item{class}{}

\item{icon}{}

\item{label}{}
}
\value{
character html fragment
}
\description{
button_modal
}

