% Generated by roxygen2 (4.0.2): do not edit by hand
\name{button_page_external}
\alias{button_page_external}
\title{button_page_external}
\usage{
button_page_external(url, label = "")
}
\arguments{
\item{url}{character, external link e.g. "https://ijlyttle.shinyapps.io/introduction"}

\item{label}{character, html for the button}
}
\value{
html fragment describing a div with a centered button
}
\description{
generate html for a button to launch
an external page in a new browser-tab
}

