% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{div_modal}
\alias{div_modal}
\title{div_modal}
\usage{
div_modal(id, div_content = "", title = "", size = "medium")
}
\arguments{
\item{id}{character, unique within page, to identify the modal panel}

\item{div_content}{character, markdown or html to use as content for modal ("")}

\item{title}{character, title to display for modal ("")}
}
\value{
html div
}
\description{
creates a button and modal window
}

