% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_caption.R
\name{img_caption}
\alias{img_caption}
\title{img_caption}
\usage{
img_caption(src, caption = "")
}
\arguments{
\item{src}{string describing location of image}

\item{caption}{string for caption}
}
\value{
string html-code sepcifying table
}
\description{
creates an aligned table with an image and a caption
}

