% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/md2html_fragment.R
\name{md2html_fragment}
\alias{md2html_fragment}
\title{md2html_fragment}
\usage{
md2html_fragment(text_markdown)
}
\arguments{
\item{text_markdown}{}
}
\value{
character, html fragment
}
\description{
create an html fragment using markdown text
}

