% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel.R
\name{panel_exercise}
\alias{panel_exercise}
\title{panel_exercise}
\usage{
panel_exercise(div_content_panel = "", title_panel = "Exercise", id_modal,
  div_content_modal = "", title_modal = "Answer")
}
\arguments{
\item{div_content_panel}{character, markdown or html to use as content for panel ("")}

\item{title_panel}{character, title to display for panel ("Exercise")}

\item{id_modal}{character, unique within page, to identify the modal panel}

\item{div_content_modal}{character, markdown or html to use as content for modal ("")}

\item{title_modal}{character, title to display for modal ("Answer")}
}
\value{
html for panel and modal
}
\description{
builds an html panel, with a possibility for modal/modal botton at bottom
}
\details{
At the bottom of the panel, there will be a button to activate a modal.
}

