% Generated by roxygen2 (4.0.2): do not edit by hand
\name{panel_html}
\alias{panel_html}
\title{panel_html}
\usage{
panel_html(title = "", body_html = "", class = "panel panel-default")
}
\arguments{
\item{title}{character, title for panel}

\item{body_html}{character, html for body}

\item{class}{character, name of bootstrap panel class}
}
\value{
html fragment describing a bootstrap panel
}
\description{
Creates a panel
}

