% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel.R
\name{panel_markdown}
\alias{panel_markdown}
\title{panel_markdown}
\usage{
panel_markdown(title = "", body_markdown = "",
  class = "panel panel-default")
}
\arguments{
\item{title}{character, title for panel}

\item{body_markdown}{character, markdown for body}

\item{class}{character, name of bootstrap panel class}
}
\value{
html fragment describing a bootstrap panel
}
\description{
Creates a well panel with a centered title an html body,
generated using markdown
}

