% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_pallette.R
\name{se_palette}
\alias{se_palette}
\title{se_palette}
\usage{
se_palette(name = NULL)
}
\arguments{
\item{name}{vector of colors to return (default NULL to return all)}
}
\value{
character vector
}
\description{
function to return colors of the Schneider Electric palette
}

