% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/well.R
\name{well_html}
\alias{well_html}
\title{well_html}
\usage{
well_html(title = "", body_html = "")
}
\arguments{
\item{title}{character, title for panel}

\item{body_html}{character, html for body}
}
\value{
html fragment describing a bootstrap well-panel
}
\description{
Creates a well panel with a centered title an html body
}

