% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_markdown}
\alias{well_markdown}
\title{well_markdown}
\usage{
well_markdown(title = "", body_markdown = "")
}
\arguments{
\item{title}{character, title for panel}

\item{body_markdown}{character, markdown for body}
}
\value{
html fragment describing a bootstrap well-panel
}
\description{
Creates a well panel with a centered title an html body,
generated using markdown
}

