% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_modal.R
\name{bs_modal}
\alias{bs_modal}
\title{bs_modal}
\usage{
bs_modal(id, title, trigger, ...)
}
\arguments{
\item{id}{blah}

\item{title}{blah}

\item{trigger}{blah}

\item{...}{}
}
\value{
character, html fragment
}
\description{
create html for bootstrap modal - based on ebailey78/shinyBS bsModal
}

