% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bs_modal.R
\name{button_modal}
\alias{button_modal}
\title{button_modal}
\usage{
button_modal(id, idModal, icon = NULL, label = "")
}
\arguments{
\item{id}{}

\item{idModal}{}

\item{icon}{}

\item{label}{}

\item{class}{}
}
\value{
character html fragment
}
\description{
button_modal
}

