% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_modal.R
\name{button_modal}
\alias{button_modal}
\title{button_modal}
\usage{
button_modal(id, idModal, icon = NULL, label = "")
}
\arguments{
\item{id}{blah}

\item{idModal}{blah}

\item{icon}{blah}

\item{label}{blah}

\item{class}{blah}
}
\value{
character html fragment
}
\description{
button_modal
}

