% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_keynote.R
\name{ggsave_black}
\alias{ggsave_black}
\title{ggsave_black}
\usage{
ggsave_black(file_name, plot)
}
\arguments{
\item{file_name}{blah}

\item{plot}{blah}
}
\description{
save ggplot using black-background theme
}

