% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel.R
\name{panel_html}
\alias{panel_html}
\title{panel_html}
\usage{
panel_html(div_content, title = "", class = "panel panel-default")
}
\arguments{
\item{div_content}{div, html for body}

\item{title}{character, title for panel}

\item{class}{character, name of bootstrap panel class}
}
\value{
html fragment describing a bootstrap panel
}
\description{
Creates a panel
}

