% Generated by roxygen2 (4.0.2): do not edit by hand
\name{panel_html_exercise}
\alias{panel_html_exercise}
\title{panel_html_exercise}
\usage{
panel_html_exercise(title = "", body_html = "",
  class = "panel panel-info")
}
\arguments{
\item{title}{character, title for panel}

\item{body_html}{character, html for body}
}
\value{
html fragment describing a bootstrap panel
}
\description{
Wraps panel_html, setting the class of the panel
}

