% Generated by roxygen2 (4.0.2): do not edit by hand
\name{panel_markdown_section}
\alias{panel_markdown_section}
\title{panel_markdown_section}
\usage{
panel_markdown_section(title = "", body_markdown = "")
}
\arguments{
\item{title}{character, title for panel}

\item{body_markdown}{character, markdown for body}
}
\value{
html fragment describing a bootstrap panel
}
\description{
Wraps panel_markdown, setting the class of the panel
}

