% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel.R
\name{panel_section}
\alias{panel_section}
\title{panel_section}
\usage{
panel_section(div_content = "", title = "")
}
\arguments{
\item{div_content}{character, markdown or html to use as content for panel ("")}

\item{title}{character, title to display for panel ("")}
}
\value{
html for panel
}
\description{
builds an html panel, with a possibility for modal/modal botton at bottom
}
\details{
At the bottom of the panel, there will be a button to activate a modal.
}

