% Generated by roxygen2 (4.0.1): do not edit by hand
\name{well_markdown}
\alias{well_markdown}
\title{well_markdown}
\usage{
well_markdown(title = "", body_markdown = "")
}
\arguments{
\item{title}{}

\item{body_markdown}{}
}
\value{
html fragment describing a bootstrap well-panel
}
\description{
Creates a well panel with a centered title an html body,
generated using markdown
}

